issue-731: Renamed widgets to conform iOS naming
