+++ android/app/src/main/java/fi/fmi/mobileweather/LargeForecastWidgetProvider.java
@@ -1,22 +1,16 @@
 package fi.fmi.mobileweather;
 
-import static android.appwidget.AppWidgetManager.ACTION_APPWIDGET_UPDATE;
-import static fi.fmi.mobileweather.WidgetNotification.ACTION_APPWIDGET_AUTO_UPDATE;
-
-import android.appwidget.AppWidgetManager;
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-
+import android.text.Html;
 import android.util.Log;
 import android.widget.RemoteViews;
-import android.view.View;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.Iterator;
+import java.util.Locale;
 
 import static fi.fmi.mobileweather.PrefKey.WIDGET_UI_UPDATED;
 
@@ -26,10 +20,8 @@ public class LargeForecastWidgetProvider extends BaseWidgetProvider {
         return R.layout.large_forecast_widget_layout;
     }
 
-    // populate widget with data
-
     private double getTimestepCount(int widgetWidth) {
-        final int columnWidth = 52;
+        final int columnWidth = 46;
         final int margins = 32;
 
         return Math.floor((widgetWidth - margins)/columnWidth);
@@ -39,9 +31,9 @@ public class LargeForecastWidgetProvider extends BaseWidgetProvider {
     protected void setWidgetData(JSONArray announcementsJson, SharedPreferencesHelper pref, WidgetInitResult widgetInitResult, int appWidgetId) {
         JSONObject forecastJson = widgetInitResult.forecastJson();
         RemoteViews widgetRemoteViews = widgetInitResult.widgetRemoteViews();
-
         final double timeStepCount = getTimestepCount(getWidgetWidthInPixels(appWidgetId));
-        Log.d("setWidgetData", "widget width: "+String.valueOf(getWidgetWidthInPixels(appWidgetId)));
+
+        Log.d("widgetWidth", String.valueOf(getWidgetWidthInPixels(appWidgetId)));
 
         try {
             // Get the keys of the JSONObject
@@ -59,38 +51,56 @@ public class LargeForecastWidgetProvider extends BaseWidgetProvider {
 
             // find first epoch time which is in future
             int firstFutureTimeIndex = getFirstFutureTimeIndex(data);
-            // if no future time found or less than 5 future times available, do not continue
-            if (firstFutureTimeIndex == -1 || data.length() < (firstFutureTimeIndex + 5)) {
+            // if no future time found or less than 6 future times available, do not continue
+            if (firstFutureTimeIndex == -1 || data.length() < (firstFutureTimeIndex + 6)) {
                 // throw new Exception("No future time found or less than 5 future times available");
-                throw new Exception("No future time found or less than 5 future times available");
+                throw new Exception("No future time found or less than 6 future times available");
             }
 
-            widgetRemoteViews.removeAllViews(R.id.hourForecastRowLayout);
+            widgetRemoteViews.removeAllViews(R.id.forecastContainer);
 
-            // handle the first 5 JsonObjects with future time
+            // handle the first 6 JsonObjects with future time
             for (int i = firstFutureTimeIndex; i < (firstFutureTimeIndex + timeStepCount); i++) {
                 JSONObject forecast = data.getJSONObject(i);
 
-                // if first future index
+                // if first future index set main part of the widget
                 if (i == firstFutureTimeIndex) {
                     // set the location name and region
                     String name = forecast.getString("name");
                     String region = forecast.getString("region");
                     widgetRemoteViews.setTextViewText(R.id.locationNameTextView, name + ", ");
                     widgetRemoteViews.setTextViewText(R.id.locationRegionTextView, region);
-                }
 
-                RemoteViews timeStep = new RemoteViews(context.getPackageName(), R.layout.forecast_timestep);
+                    String localTime = forecast.getString("localtime");
+                    String formattedTime = getFormattedWeatherTime(localTime);
+                    widgetRemoteViews.setTextViewText(R.id.timeTextView, formattedTime);
+
+                    String temperature = forecast.getString("temperature");
+                    int weatherSymbol = forecast.getInt("smartSymbol");
+
+                    widgetRemoteViews.setTextViewText(R.id.temperatureTextView, addPlusIfNeeded(temperature) + "°");
+
+                    int drawableResId = context.getResources().getIdentifier("s_" + weatherSymbol, "drawable", context.getPackageName());
+                    widgetRemoteViews.setImageViewResource(R.id.weatherIconImageView, drawableResId);
+                    widgetRemoteViews.setContentDescription(R.id.weatherIconImageView, getSymbolTranslation(weatherSymbol));
+
+                    // next iteration in loop
+                    continue;
+                }
+                RemoteViews timeStep = new RemoteViews(context.getPackageName(), R.layout.medium_forecast_timestep);
 
                 // time at the selected location
                 String localTime = forecast.getString("localtime");
                 String temperature = forecast.getString("temperature");
                 int weatherSymbol = forecast.getInt("smartSymbol");
 
+                // j = weather row layout index
+                int j = i - 1;
+
                 // ** set the time, temperature and weather icon
 
                 String formattedTime = getFormattedWeatherTime(localTime);
-                timeStep.setTextViewText(R.id.timeStepTimeTextView, formattedTime);
+                timeStep.setTextViewText(R.id.timeTextView, formattedTime);
 
                 temperature = addPlusIfNeeded(temperature);
                 timeStep.setTextViewText(R.id.temperatureTextView, temperature + "°");
@@ -99,12 +109,31 @@ public class LargeForecastWidgetProvider extends BaseWidgetProvider {
                 timeStep.setImageViewResource(R.id.weatherIconImageView, drawableResId);
                 timeStep.setContentDescription(R.id.weatherIconImageView, getSymbolTranslation(weatherSymbol));
 
-                if (i == firstFutureTimeIndex + timeStepCount - 1) {
-                    timeStep.setViewVisibility(R.id.forecastBorder, View.GONE);
-                }
-
-                widgetRemoteViews.addView(R.id.hourForecastRowLayout, timeStep);
+                widgetRemoteViews.addView(R.id.forecastContainer, timeStep);
             }
+
+            // Get the current time
+            Date currentTime = new Date();
+            // Format the time as "HH:mm"
+            SimpleDateFormat formatter = new SimpleDateFormat("HH:mm", Locale.getDefault());
+            String formattedTime = formatter.format(currentTime);
+
+            String htmlString = context.getString(R.string.updated)
+                    + " "
+                    + "<b>"
+                    // add time of now in HH:mm format
+                    + formattedTime
+                    + "</b>";
+
+            // Convert the HTML string to a CharSequence
+            CharSequence formattedText;
+            if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
+                formattedText = Html.fromHtml(htmlString, Html.FROM_HTML_MODE_LEGACY);
+            } else {
+                formattedText = Html.fromHtml(htmlString);
+            }
+
+            widgetRemoteViews.setTextViewText(R.id.updateTimeTextView, formattedText);
 
             // Crisis view
             showCrisisViewIfNeeded(announcementsJson, widgetRemoteViews, pref, true);
@@ -125,4 +154,5 @@ public class LargeForecastWidgetProvider extends BaseWidgetProvider {
 
         appWidgetManager.updateAppWidget(appWidgetId, widgetRemoteViews);
     }
+
 }
