--- android/app/src/main/java/fi/fmi/mobileweather/MaxForecastWidgetConfigurationActivity.java
@@ -1,8 +1,0 @@
-package fi.fmi.mobileweather;
-
-public class MaxForecastWidgetConfigurationActivity extends BaseWidgetConfigurationActivity {
-    @Override
-    protected Class<?> getWidgetProviderClass() {
-        return MaxForecastWidgetProvider.class;
-    }
-}
