+++ android/app/src/main/java/fi/fmi/mobileweather/MediumForecastWidgetConfigurationActivity.java
@@ -1,0 +1,8 @@
+package fi.fmi.mobileweather;
+
+public class MediumForecastWidgetConfigurationActivity extends BaseWidgetConfigurationActivity {
+    @Override
+    protected Class<?> getWidgetProviderClass() {
+        return MediumForecastWidgetProvider.class;
+    }
+}
