+++ android/app/src/main/java/fi/fmi/mobileweather/MediumForecastWidgetProvider.java
@@ -1,32 +1,26 @@
 package fi.fmi.mobileweather;
 
-import android.appwidget.AppWidgetManager;
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-
-import android.text.Html;
 import android.util.Log;
 import android.widget.RemoteViews;
+import android.view.View;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.Iterator;
-import java.util.Locale;
 
 import static fi.fmi.mobileweather.PrefKey.WIDGET_UI_UPDATED;
 
-public class MaxForecastWidgetProvider extends BaseWidgetProvider {
+public class MediumForecastWidgetProvider extends BaseWidgetProvider {
     @Override
     protected int getLayoutResourceId() {
-        return R.layout.max_forecast_widget_layout;
+        return R.layout.medium_forecast_widget_layout;
     }
 
+    // populate widget with data
+
     private double getTimestepCount(int widgetWidth) {
-        final int columnWidth = 46;
+        final int columnWidth = 52;
         final int margins = 32;
 
         return Math.floor((widgetWidth - margins)/columnWidth);
@@ -36,9 +30,9 @@ public class MaxForecastWidgetProvider extends BaseWidgetProvider {
     protected void setWidgetData(JSONArray announcementsJson, SharedPreferencesHelper pref, WidgetInitResult widgetInitResult, int appWidgetId) {
         JSONObject forecastJson = widgetInitResult.forecastJson();
         RemoteViews widgetRemoteViews = widgetInitResult.widgetRemoteViews();
-        final double timeStepCount = getTimestepCount(getWidgetWidthInPixels(appWidgetId));
 
-        Log.d("widgetWidth", String.valueOf(getWidgetWidthInPixels(appWidgetId)));
+        final double timeStepCount = getTimestepCount(getWidgetWidthInPixels(appWidgetId));
+        Log.d("setWidgetData", "widget width: "+String.valueOf(getWidgetWidthInPixels(appWidgetId)));
 
         try {
             // Get the keys of the JSONObject
@@ -56,56 +50,38 @@ public class MaxForecastWidgetProvider extends BaseWidgetProvider {
 
             // find first epoch time which is in future
             int firstFutureTimeIndex = getFirstFutureTimeIndex(data);
-            // if no future time found or less than 6 future times available, do not continue
-            if (firstFutureTimeIndex == -1 || data.length() < (firstFutureTimeIndex + 6)) {
+            // if no future time found or less than 5 future times available, do not continue
+            if (firstFutureTimeIndex == -1 || data.length() < (firstFutureTimeIndex + 5)) {
                 // throw new Exception("No future time found or less than 5 future times available");
-                throw new Exception("No future time found or less than 6 future times available");
+                throw new Exception("No future time found or less than 5 future times available");
             }
 
-            widgetRemoteViews.removeAllViews(R.id.forecastContainer);
+            widgetRemoteViews.removeAllViews(R.id.hourForecastRowLayout);
 
-            // handle the first 6 JsonObjects with future time
+            // handle the first 5 JsonObjects with future time
             for (int i = firstFutureTimeIndex; i < (firstFutureTimeIndex + timeStepCount); i++) {
                 JSONObject forecast = data.getJSONObject(i);
 
-                // if first future index set main part of the widget
+                // if first future index
                 if (i == firstFutureTimeIndex) {
                     // set the location name and region
                     String name = forecast.getString("name");
                     String region = forecast.getString("region");
                     widgetRemoteViews.setTextViewText(R.id.locationNameTextView, name + ", ");
                     widgetRemoteViews.setTextViewText(R.id.locationRegionTextView, region);
-
-                    String localTime = forecast.getString("localtime");
-                    String formattedTime = getFormattedWeatherTime(localTime);
-                    widgetRemoteViews.setTextViewText(R.id.timeTextView, formattedTime);
-
-                    String temperature = forecast.getString("temperature");
-                    int weatherSymbol = forecast.getInt("smartSymbol");
-
-                    widgetRemoteViews.setTextViewText(R.id.temperatureTextView, addPlusIfNeeded(temperature) + "°");
-
-                    int drawableResId = context.getResources().getIdentifier("s_" + weatherSymbol, "drawable", context.getPackageName());
-                    widgetRemoteViews.setImageViewResource(R.id.weatherIconImageView, drawableResId);
-                    widgetRemoteViews.setContentDescription(R.id.weatherIconImageView, getSymbolTranslation(weatherSymbol));
-
-                    // next iteration in loop
-                    continue;
                 }
-                RemoteViews timeStep = new RemoteViews(context.getPackageName(), R.layout.large_forecast_timestep);
+
+                RemoteViews timeStep = new RemoteViews(context.getPackageName(), R.layout.forecast_timestep);
 
                 // time at the selected location
                 String localTime = forecast.getString("localtime");
                 String temperature = forecast.getString("temperature");
                 int weatherSymbol = forecast.getInt("smartSymbol");
 
-                // j = weather row layout index
-                int j = i - 1;
-
                 // ** set the time, temperature and weather icon
 
                 String formattedTime = getFormattedWeatherTime(localTime);
-                timeStep.setTextViewText(R.id.timeTextView, formattedTime);
+                timeStep.setTextViewText(R.id.timeStepTimeTextView, formattedTime);
 
                 temperature = addPlusIfNeeded(temperature);
                 timeStep.setTextViewText(R.id.temperatureTextView, temperature + "°");
@@ -114,31 +90,12 @@ public class MaxForecastWidgetProvider extends BaseWidgetProvider {
                 timeStep.setImageViewResource(R.id.weatherIconImageView, drawableResId);
                 timeStep.setContentDescription(R.id.weatherIconImageView, getSymbolTranslation(weatherSymbol));
 
-                widgetRemoteViews.addView(R.id.forecastContainer, timeStep);
+                if (i == firstFutureTimeIndex + timeStepCount - 1) {
+                    timeStep.setViewVisibility(R.id.forecastBorder, View.GONE);
+                }
+
+                widgetRemoteViews.addView(R.id.hourForecastRowLayout, timeStep);
             }
-
-            // Get the current time
-            Date currentTime = new Date();
-            // Format the time as "HH:mm"
-            SimpleDateFormat formatter = new SimpleDateFormat("HH:mm", Locale.getDefault());
-            String formattedTime = formatter.format(currentTime);
-
-            String htmlString = context.getString(R.string.updated)
-                    + " "
-                    + "<b>"
-                    // add time of now in HH:mm format
-                    + formattedTime
-                    + "</b>";
-
-            // Convert the HTML string to a CharSequence
-            CharSequence formattedText;
-            if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
-                formattedText = Html.fromHtml(htmlString, Html.FROM_HTML_MODE_LEGACY);
-            } else {
-                formattedText = Html.fromHtml(htmlString);
-            }
-
-            widgetRemoteViews.setTextViewText(R.id.updateTimeTextView, formattedText);
 
             // Crisis view
             showCrisisViewIfNeeded(announcementsJson, widgetRemoteViews, pref, true);
@@ -159,5 +116,4 @@ public class MaxForecastWidgetProvider extends BaseWidgetProvider {
 
         appWidgetManager.updateAppWidget(appWidgetId, widgetRemoteViews);
     }
-
 }
