+++ android/app/src/main/res/layout/large_forecast_widget_layout.xml
@@ -1,27 +1,26 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mainLinearLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/widgetBackground"
     android:orientation="vertical"
-    >
+    android:padding="16dp">
 
     <LinearLayout
         android:id="@+id/normalLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical">
+        android:orientation="vertical"
+        tools:visibility="visible">
 
         <LinearLayout
             android:id="@+id/locationLayout"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:gravity="top"
-            android:layout_marginHorizontal="16dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="4dp">
+            android:layout_gravity="center_horizontal"
+            android:orientation="horizontal">
 
             <TextView
                 android:id="@+id/locationNameTextView"
@@ -44,19 +43,74 @@
 
         </LinearLayout>
 
-        <LinearLayout
-            android:id="@+id/hourForecastRowLayout"
+        <TextView
+            android:id="@+id/timeTextView"
+            style="@style/currentWeatherTime"
             android:layout_width="wrap_content"
-            android:layout_height="60dp"
-            android:layout_marginHorizontal="12dp"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal"
+            android:gravity="center_horizontal"
+            android:textSize="15sp"
+            tools:text="klo 11:00" />
+
+        <LinearLayout
+            android:id="@+id/weatherLayout"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal"
+            android:layout_marginTop="3dp"
+            android:layout_marginBottom="3dp"
             android:orientation="horizontal">
+
+            <ImageView
+                android:id="@+id/weatherIconImageView"
+                android:layout_width="46dp"
+                android:layout_height="46dp"
+                android:layout_marginEnd="18dp"
+                tools:src="@drawable/s_32" />
+
+            <TextView
+                android:id="@+id/temperatureTextView"
+                style="@style/temperature"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_vertical"
+                tools:text="+12" />
+
+            <TextView
+                android:id="@+id/temperatureUnitTextView"
+                style="@style/temperatureUnit"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="6dp"
+                tools:text="°C" />
+
         </LinearLayout>
+
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            >
+            <include layout="@layout/medium_forecast_background" />
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+                <include layout="@layout/forecast_symbol_column" />
+                <LinearLayout
+                    android:id="@+id/forecastContainer"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_alignParentEnd="true"
+                    android:orientation="horizontal"
+                    android:layout_marginHorizontal="6dp" />
+            </RelativeLayout>
+        </FrameLayout>
 
         <LinearLayout
             android:id="@+id/crisisViewContainer"
             android:layout_height="wrap_content"
             android:layout_width="match_parent"
-            android:paddingHorizontal="16dp"
             android:orientation="vertical" />
 
     </LinearLayout>
@@ -65,10 +119,8 @@
         android:id="@+id/errorLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:gravity="center_vertical"
-        android:orientation="vertical"
-        android:visibility="gone"
-        tools:visibility="visible">
-        <include layout="@layout/medium_errorview" />
+        android:visibility="gone">
+        <include layout="@layout/large_errorview" />
     </LinearLayout>
+
 </LinearLayout>
