+++ android/app/src/main/res/layout/medium_forecast_background.xml
