+++ android/app/src/main/res/xml-v31/large_widget_provider_info.xml
@@ -2,11 +2,11 @@
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
     android:configure="fi.fmi.mobileweather.LargeForecastWidgetConfigurationActivity"
     android:initialLayout="@layout/large_forecast_widget_layout"
-    android:targetCellWidth="4"
-    android:targetCellHeight="1"
-    android:previewImage="@drawable/medium_forecast_review"
-    android:minResizeWidth="220dp"
+    android:previewImage="@drawable/large_forecast_review"
     android:resizeMode="horizontal"
+    android:targetCellWidth="4"
+    android:targetCellHeight="2"
+    android:minResizeWidth="250dp"
     android:updatePeriodMillis="0"
     android:widgetCategory="home_screen">
 </appwidget-provider>
