--- android/app/src/main/res/xml-v31/max_widget_provider_info.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:configure="fi.fmi.mobileweather.MaxForecastWidgetConfigurationActivity"
-    android:initialLayout="@layout/max_forecast_widget_layout"
-    android:previewImage="@drawable/large_forecast_review"
-    android:resizeMode="horizontal"
-    android:targetCellWidth="4"
-    android:targetCellHeight="2"
-    android:minResizeWidth="250dp"
-    android:updatePeriodMillis="0"
-    android:widgetCategory="home_screen">
-</appwidget-provider>
