+++ android/app/src/main/res/xml-v31/medium_widget_provider_info.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
+    android:configure="fi.fmi.mobileweather.MediumForecastWidgetConfigurationActivity"
+    android:initialLayout="@layout/medium_forecast_widget_layout"
+    android:targetCellWidth="4"
+    android:targetCellHeight="1"
+    android:previewImage="@drawable/medium_forecast_review"
+    android:minResizeWidth="220dp"
+    android:resizeMode="horizontal"
+    android:updatePeriodMillis="0"
+    android:widgetCategory="home_screen">
+</appwidget-provider>
